hash code type use size_t

+++ servant/libservant/ServantProxy.cpp
@@ -697,7 +697,7 @@ void ServantProxy::tars_async_ping()
 	tars_invoke_async(TARSONEWAY, "tars_ping", os, m, s, NULL);
 }
 
-ServantProxy* ServantProxy::tars_hash(uint32_t key)
+ServantProxy* ServantProxy::tars_hash(size_t key)
 {
     ServantProxyThreadData *pSptd = ServantProxyThreadData::getData();
 
@@ -717,7 +717,7 @@ ServantProxy* ServantProxy::tars_hash(uint32_t key)
 //	return this;
 //}
 
-ServantProxy* ServantProxy::tars_consistent_hash(uint32_t key)
+ServantProxy* ServantProxy::tars_consistent_hash(size_t key)
 {
     ServantProxyThreadData *pSptd = ServantProxyThreadData::getData();
